separate service to multiple services by logic
