+++ src/main/kotlin/ru/hse/dynamomock/service/AWSDynamoDBMockService.kt
@@ -3,455 +3,39 @@ package ru.hse.dynamomock.service
 import com.github.doyaaaaaken.kotlincsv.dsl.csvReader
 import ru.hse.dynamomock.db.DataStorageLayer
 import ru.hse.dynamomock.exception.AWSMockCSVException
-import ru.hse.dynamomock.exception.dynamoException
-import ru.hse.dynamomock.exception.dynamoRequires
 import ru.hse.dynamomock.model.*
-import ru.hse.dynamomock.model.query.*
-import software.amazon.awssdk.core.util.DefaultSdkAutoConstructList
-import software.amazon.awssdk.core.util.DefaultSdkAutoConstructMap
 import software.amazon.awssdk.services.dynamodb.model.*
-import java.util.*
 import java.io.File
 import java.io.InputStream
 
-class AWSDynamoDBMockService(private val storage: DataStorageLayer) {
-    private val tablesMetadata = mutableMapOf<String, TableMetadata>()
-
-    private fun toKey(keyName: String, attributeValue: AttributeValue): Pair<DynamoType, Key> =
-        if (attributeValue.s() != null) {
-            DynamoType.S to StringKey(keyName, attributeValue.s())
-        } else if (attributeValue.n() != null) {
-            DynamoType.N to NumKey(keyName, attributeValue.n().toBigDecimal())
-        } else if (attributeValue.b() != null) {
-            DynamoType.B to StringKey(keyName, Base64.getEncoder().encodeToString(attributeValue.b().asByteArray()))
-        } else {
-            throw dynamoException("Member must satisfy enum value set: [B, N, S]")
-        }
+class AWSDynamoDBMockService(
+    private val storage: DataStorageLayer,
+    private val ddlService: DDLService,
+    private val dmlService: DMLService,
+    private val selectService: SelectService
+) {
 
     fun close() = storage.close()
 
-    fun createTable(request: CreateTableRequest): TableDescription {
-        dynamoRequires(request.tableName() !in tablesMetadata) {
-            "Table ${request.tableName()} already exists. Cannot create."
-        }
-        return request.toTableMetadata().also {
-            storage.createTable(it)
-            tablesMetadata[it.tableName] = it
-        }.toTableDescription()
-    }
+    fun createTable(request: CreateTableRequest): TableDescription = ddlService.createTable(request)
 
-    fun deleteTable(request: DeleteTableRequest): TableDescription {
-        val name = request.tableName()
-        dynamoRequires(name in tablesMetadata) {
-            "Cannot delete non-existent table."
-        }
-        storage.deleteTable(name)
-        return tablesMetadata.remove(name)!!.toTableDescription()
-    }
+    fun deleteTable(request: DeleteTableRequest): TableDescription = ddlService.deleteTable(request)
 
-    fun describeTable(request: DescribeTableRequest): TableDescription {
-        val name = request.tableName()
-        dynamoRequires(name in tablesMetadata) {
-            "Cannot describe non-existent table $name."
-        }
-        return tablesMetadata.getValue(name).toTableDescription()
-    }
+    fun describeTable(request: DescribeTableRequest): TableDescription = ddlService.describeTable(request)
 
-    fun query(request: QueryRequest): QueryResponse = request.toSelectRequest().select().toQueryResponse()
+    fun query(request: QueryRequest): QueryResponse = selectService.query(request)
 
-    fun scan(request: ScanRequest): ScanResponse = request.toSelectRequest().select().toScanResponse()
+    fun scan(request: ScanRequest): ScanResponse = selectService.scan(request)
 
-    fun putItem(request: PutItemRequest): PutItemResponse {
-        val tableName = request.tableName()
+    fun putItem(request: PutItemRequest): PutItemResponse = dmlService.putItem(request)
 
-        val (partitionKey, sortKey) = getRequestMetadata(tableName, request.item())
+    fun getItem(request: GetItemRequest): GetItemResponse = dmlService.getItem(request)
 
-        val itemsList = request.item().map { (k, v) -> AttributeInfo(k, v.toAttributeTypeInfo()) }
+    fun deleteItem(request: DeleteItemRequest): DeleteItemResponse = dmlService.deleteItem(request)
 
-        val previousItem = storage.getItem(DBGetItemRequest(tableName, partitionKey, sortKey))?.associate {
-            it.name to it.type.toAttributeValue()
-        }
+    fun updateItem(request: UpdateItemRequest): UpdateItemResponse = dmlService.updateItem(request)
 
-        if (previousItem != null) {
-            storage.updateItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-        } else {
-            storage.putItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-        }
-
-        val attributes = getAttributesFromReturnValues(request.returnValues(), previousItem)
-
-        return PutItemResponse.builder()
-            .attributes(attributes)
-            .build()
-    }
-
-    fun getItem(request: GetItemRequest): GetItemResponse {
-        val tableName = request.tableName()
-        checkNumOfKeys(tableName, request.key())
-        val (partitionKey, sortKey) = getRequestMetadata(tableName, request.key())
-
-        val response = storage.getItem(DBGetItemRequest(tableName, partitionKey, sortKey))
-
-        val item = response?.filter { it.name in request.attributesToGet() }?.associate {
-            it.name to it.type.toAttributeValue()
-        }
-
-        return GetItemResponse.builder()
-            .item(item)
-            .build()
-    }
-
-    fun deleteItem(request: DeleteItemRequest): DeleteItemResponse {
-        val tableName = request.tableName()
-        checkNumOfKeys(tableName, request.key())
-        val (partitionKey, sortKey) = getRequestMetadata(tableName, request.key())
-
-        val previousItem = storage.getItem(DBGetItemRequest(tableName, partitionKey, sortKey))?.associate {
-            it.name to it.type.toAttributeValue()
-        }
-
-        if (previousItem != null) {
-            storage.deleteItem(DBDeleteItemRequest(tableName, partitionKey, sortKey))
-        }
-
-        val attributes = getAttributesFromReturnValues(request.returnValues(), previousItem)
-
-        return DeleteItemResponse.builder()
-            .attributes(attributes)
-            .build()
-    }
-
-    private fun buildItem(
-        keys: Map<String, AttributeValue>,
-        columnName: String,
-        attributeValue: AttributeValue
-    ): Map<String, AttributeValue> = mapOf(columnName to attributeValue) + keys
-
-
-    private fun checkAttributeValue(av: AttributeValue): DynamoType = when {
-            av.n() != null -> DynamoType.N
-            av.s() != null -> DynamoType.S
-            av.b() != null -> DynamoType.B
-            av.hasSs() -> DynamoType.SS
-            av.hasNs() -> DynamoType.NS
-            av.hasBs() -> DynamoType.BS
-            av.hasL() -> DynamoType.L
-            av.hasM() -> DynamoType.M
-            else -> throw dynamoException("Supplied AttributeValue is empty, must contain exactly one of the supported datatypes")
-        }
-
-    private fun checkAttributesForDelete(first: AttributeValue, second: AttributeValue): DynamoType {
-        val firstType = checkAttributeValue(first)
-        val secondType = checkAttributeValue(second)
-        if (secondType !in listOf(DynamoType.NS, DynamoType.SS, DynamoType.BS)) {
-            val errorString =
-                "One or more parameter values were invalid: DELETE action with value is not supported for the type " +
-                    when (secondType) {
-                        DynamoType.N -> "NUMBER"
-                        DynamoType.S -> "STRING"
-                        DynamoType.B -> "BINARY"
-                        DynamoType.BOOL -> "BOOLEAN"
-                        DynamoType.NULL -> "NULL"
-                        DynamoType.L -> "LIST"
-                        DynamoType.M -> "MAP"
-                        else -> ""
-                    }
-            throw dynamoException(errorString)
-        }
-        if (firstType != secondType) {
-            throw dynamoException("Type mismatch for attribute to update")
-        }
-        return secondType
-    }
-
-    private fun putInUpdate(
-        tableName: String, partitionKey: Key, sortKey: Key?,
-        columnName: String, av: AttributeValueUpdate, keys: Map<String, AttributeValue>
-    ) {
-        val previousItem = storage.getItem(DBGetItemRequest(tableName, partitionKey, sortKey))?.associate {
-            it.name to it.type.toAttributeValue()
-        }
-        if (previousItem != null) {
-            // gel all columns
-            val itemsList = previousItem.map { (k, v) ->
-                AttributeInfo(k, v.toAttributeTypeInfo())
-            }.toMutableList()
-            val newAttributeTypeInfo = av.value().toAttributeTypeInfo()
-
-            // check if columnName is there
-            val previousAttributeValue = previousItem[columnName]
-            if (previousAttributeValue != null) {
-                val oldAttributeTypeInfo = previousAttributeValue.toAttributeTypeInfo()
-                itemsList.remove(AttributeInfo(columnName, oldAttributeTypeInfo))
-            }
-            itemsList.add(AttributeInfo(columnName, newAttributeTypeInfo))
-            storage.updateItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-        } else {
-            val itemsList = buildItem(keys, columnName, av.value()).map { (k, v) ->
-                AttributeInfo(
-                    k,
-                    v.toAttributeTypeInfo()
-                )
-            }
-            storage.putItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-        }
-    }
-
-    private fun deleteInUpdate(
-        tableName: String, partitionKey: Key, sortKey: Key?,
-        columnName: String, av: AttributeValueUpdate, keys: Map<String, AttributeValue>
-    ) {
-        val previousItem = storage.getItem(DBGetItemRequest(tableName, partitionKey, sortKey))?.associate {
-            it.name to it.type.toAttributeValue()
-        } ?: throw dynamoException("One of the required keys was not given a value")
-        if (av.value() == null) {
-            val newItem = previousItem.toMutableMap()
-            newItem.remove(columnName)
-            val itemsList = newItem.map { (k, v) -> AttributeInfo(k, v.toAttributeTypeInfo()) }
-            storage.updateItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-        } else {
-            val previousItemAttribute = previousItem[columnName]
-            if (previousItemAttribute != null) {
-                val setType = checkAttributesForDelete(previousItemAttribute, av.value())
-                val newItem = previousItem.toMutableMap()
-                val diffSet = when (setType) {
-                    DynamoType.SS -> {
-                        dynamoRequires(av.value().ss().isNotEmpty()) {
-                            "One or more parameter values were invalid: An string set may not be empty"
-                        }
-                        val diffSet = previousItemAttribute.ss() - av.value().ss()
-                        newItem[columnName] = AttributeValue.builder()
-                            .ss(diffSet)
-                            .build()
-                        diffSet
-                    }
-                    DynamoType.NS -> {
-                        dynamoRequires(av.value().ns().isNotEmpty()) {
-                            "One or more parameter values were invalid: An number set may not be empty"
-                        }
-                        val diffSet = previousItemAttribute.ns() - av.value().ns()
-                        newItem[columnName] = AttributeValue.builder()
-                            .ns(diffSet)
-                            .build()
-                        diffSet
-                    }
-                    else -> {
-                        dynamoRequires(av.value().bs().isNotEmpty()) {
-                            "One or more parameter values were invalid: An binary set may not be empty"
-                        }
-                        val diffSet = previousItemAttribute.bs() - av.value().bs()
-                        newItem[columnName] = AttributeValue.builder()
-                            .bs(diffSet)
-                            .build()
-                        diffSet
-                    }
-                }
-
-                if (diffSet.isEmpty()) {
-                    deleteItem(
-                        DeleteItemRequest.builder()
-                            .key(keys)
-                            .tableName(tableName)
-                            .build()
-                    )
-                } else {
-                    val itemsList = newItem.map { (k, v) -> AttributeInfo(k, v.toAttributeTypeInfo()) }
-                    storage.updateItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-                }
-            }
-        }
-    }
-
-    private fun checkAttributesForAdd(first: AttributeValue, second: AttributeValue): DynamoType {
-        val firstType = checkAttributeValue(first)
-        val secondType = checkAttributeValue(second)
-        if (secondType !in listOf(DynamoType.N, DynamoType.NS, DynamoType.SS, DynamoType.BS)) {
-            val errorString =
-                "One or more parameter values were invalid: ADD action with value is not supported for the type " +
-                    when (secondType) {
-                        DynamoType.S -> "STRING"
-                        DynamoType.B -> "BINARY"
-                        DynamoType.BOOL -> "BOOLEAN"
-                        DynamoType.NULL -> "NULL"
-                        DynamoType.L -> "LIST"
-                        DynamoType.M -> "MAP"
-                        else -> ""
-                    }
-            throw dynamoException(errorString)
-        }
-        if (firstType != secondType) {
-            throw dynamoException("Type mismatch for attribute to update")
-        }
-        return secondType
-    }
-
-    private fun addInUpdate(
-        tableName: String, partitionKey: Key, sortKey: Key?,
-        columnName: String, av: AttributeValueUpdate, keys: Map<String, AttributeValue>
-    ) {
-        val previousItem = storage.getItem(DBGetItemRequest(tableName, partitionKey, sortKey))?.associate {
-            it.name to it.type.toAttributeValue()
-        }
-        if (previousItem == null) {
-            // todo: ADD - Causes DynamoDB to create an item with the supplied primary key and number (or set of numbers)
-            // for the attribute value. The only data types allowed are Number and Number Set.
-            val itemsList = buildItem(keys, columnName, av.value()).map { (k, v) ->
-                AttributeInfo(
-                    k,
-                    v.toAttributeTypeInfo()
-                )
-            }
-            storage.putItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-        } else {
-            val previousItemAttribute = previousItem[columnName]
-            if (previousItemAttribute != null) {
-                val attrType = checkAttributesForAdd(previousItemAttribute, av.value())
-                val newItem = previousItem.toMutableMap()
-                when (attrType) {
-                    DynamoType.SS -> {
-                        val newSet = previousItemAttribute.ss() + av.value().ss()
-                        newItem[columnName] = AttributeValue.builder()
-                            .ss(newSet)
-                            .build()
-                    }
-                    DynamoType.NS -> {
-                        val newSet = previousItemAttribute.ns() + av.value().ns()
-                        newItem[columnName] = AttributeValue.builder()
-                            .ns(newSet)
-                            .build()
-                    }
-                    DynamoType.BS -> {
-                        val newSet = previousItemAttribute.bs() + av.value().bs()
-                        newItem[columnName] = AttributeValue.builder()
-                            .bs(newSet)
-                            .build()
-                    }
-                    else -> {
-                        // number
-                        val newNum = previousItemAttribute.n().toBigDecimal() + av.value().n().toBigDecimal()
-                        newItem[columnName] = AttributeValue.builder()
-                            .n(newNum.toString())
-                            .build()
-                    }
-                }
-                val itemsList = newItem.map { (k, v) -> AttributeInfo(k, v.toAttributeTypeInfo()) }
-                storage.updateItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-            } else {
-                checkAttributesForAdd(av.value(), av.value())
-                val newItem = previousItem.toMutableMap()
-                newItem[columnName] = av.value()
-                val itemsList = newItem.map { (k, v) -> AttributeInfo(k, v.toAttributeTypeInfo()) }
-                storage.updateItem(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-            }
-        }
-    }
-
-    fun updateItem(request: UpdateItemRequest): UpdateItemResponse {
-        // todo: if attribute is key in index then check type
-        val tableName = request.tableName()
-        val key = request.key()
-        checkNumOfKeys(tableName, key)
-        val (partitionKey, sortKey) = getRequestMetadata(tableName, key)
-
-        val attributeValues = request.attributeUpdates()
-        attributeValues.forEach { (columnName, av) ->
-            if (key.containsKey(columnName)) {
-                throw dynamoException(
-                    "One or more parameter values were invalid: Cannot update attribute ${columnName}." +
-                        "This attribute is part of the key"
-                )
-            }
-            if (av.action() == AttributeAction.PUT) {
-                putInUpdate(tableName, partitionKey, sortKey, columnName, av, key)
-            } else if (av.action() == AttributeAction.DELETE) {
-                deleteInUpdate(tableName, partitionKey, sortKey, columnName, av, key)
-            } else if (av.action() == AttributeAction.ADD) {
-                addInUpdate(tableName, partitionKey, sortKey, columnName, av, key)
-            }
-
-        }
-
-        return UpdateItemResponse.builder()
-            .build()
-    }
-
-    fun batchWriteItem(batchWriteItemRequest: BatchWriteItemRequest): BatchWriteItemResponse {
-        dynamoRequires(batchWriteItemRequest.hasRequestItems() && batchWriteItemRequest.requestItems().isNotEmpty()) {
-            "BatchWriteItem cannot have a null or no requests set"
-        }
-        val requestItems = batchWriteItemRequest.requestItems()
-        val putItemRequests = mutableListOf<DBPutItemRequest>()
-        val deleteItemRequests = mutableListOf<DBDeleteItemRequest>()
-
-        requestItems.keys.forEach {
-            tablesMetadata[it] ?: throw ResourceNotFoundException.builder()
-                .message("Cannot do operations on a non-existent table").build()
-        }
-
-        val batchSize = requestItems.entries.sumOf { it.value.size }
-
-        dynamoRequires(batchSize <= 25) {
-            "Too many items requested for the BatchWriteItem call"
-        }
-
-        requestItems.entries.forEach { tableRequests ->
-            val items = mutableSetOf<Map<String, AttributeValue>>()
-            val tableName = tableRequests.key
-            tableRequests.value.forEach {
-                val putRequest = it.putRequest()
-                val deleteRequest = it.deleteRequest()
-                if (putRequest == null && deleteRequest == null) {
-                    throw dynamoException(
-                        "Supplied AttributeValue has more than one datatypes set, must contain exactly one of the supported datatypes"
-                    )
-                } else if (putRequest != null && putRequest.hasItem() && (deleteRequest == null || !deleteRequest.hasKey())) {
-                    val (partitionKey, sortKey) = getRequestMetadata(tableName, putRequest.item())
-                    val keys = putRequest.item().filter { item ->
-                        item.key == partitionKey.attributeName || item.key == sortKey?.attributeName
-                    }
-                    dynamoRequires(keys !in items) {
-                        "Provided list of item keys contains duplicates"
-                    }
-
-                    items.add(keys)
-                    val itemsList = putRequest.item().map { (k, v) -> AttributeInfo(k, v.toAttributeTypeInfo()) }
-                    putItemRequests.add(DBPutItemRequest(tableName, partitionKey, sortKey, itemsList))
-                } else if (deleteRequest != null && deleteRequest.hasKey() && (putRequest == null || !putRequest.hasItem())) {
-                    val (partitionKey, sortKey) = getRequestMetadata(tableName, deleteRequest.key())
-
-                    val keys = deleteRequest.key()
-                    dynamoRequires(keys !in items) {
-                        "Provided list of item keys contains duplicates"
-                    }
-
-                    items.add(keys)
-                    deleteItemRequests.add(DBDeleteItemRequest(tableName, partitionKey, sortKey))
-                } else {
-                    throw dynamoException(
-                        "Supplied AttributeValue has more than one datatypes set, must contain exactly one of the supported datatypes"
-                    )
-                }
-            }
-
-        }
-
-        putItemRequests.forEach {
-            val previousItem =
-                storage.getItem(DBGetItemRequest(it.tableName, it.partitionKey, it.sortKey))?.associate { v ->
-                    v.name to v.type.toAttributeValue()
-                }
-
-            if (previousItem != null) {
-                storage.updateItem(it)
-            } else {
-                storage.putItem(it)
-            }
-        }
-        deleteItemRequests.forEach { storage.deleteItem(it) }
-
-        return BatchWriteItemResponse.builder()
-            .build()
-    }
+    fun batchWriteItem(request: BatchWriteItemRequest): BatchWriteItemResponse = dmlService.batchWriteItem(request)
 
     fun loadCSV(filePath: String, tableName: String) {
         loadCSV(File(filePath).inputStream(), tableName)
@@ -499,138 +83,4 @@ class AWSDynamoDBMockService(private val storage: DataStorageLayer) {
         }
     }
 
-    private fun getKeyFromMetadata(
-        keyName: String,
-        keys: Map<String, AttributeValue>,
-        attributeDefinitions: List<AttributeDefinition>
-    ): Key {
-        val keyAttributeValue = keys[keyName] ?: throw dynamoException("One of the required keys was not given a value")
-        val (keyType, key) = toKey(keyName, keyAttributeValue)
-        val expectedKeyType = attributeDefinitions.firstOrNull { it.attributeName() == keyName }
-            ?: throw dynamoException("One of the required keys was not given a value")
-        dynamoRequires(expectedKeyType.attributeTypeAsString().uppercase() == keyType.name) {
-            "Invalid attribute value type"
-        }
-        return key
-    }
-
-    private fun getSortKeyFromMetadata(
-        sortKeyName: String?,
-        keys: Map<String, AttributeValue>,
-        attributeDefinitions: List<AttributeDefinition>
-    ): Key? {
-        sortKeyName ?: return null
-
-        return getKeyFromMetadata(sortKeyName, keys, attributeDefinitions)
-    }
-
-    private fun getRequestMetadata(tableName: String, keys: Map<String, AttributeValue>): Pair<Key, Key?> {
-        val tableMetadata = tablesMetadata[tableName] ?: throw ResourceNotFoundException.builder()
-            .message("Cannot do operations on a non-existent table").build()
-
-        val partitionKeyName = tableMetadata.partitionKey
-        return getKeyFromMetadata(partitionKeyName, keys, tableMetadata.attributeDefinitions) to
-            getSortKeyFromMetadata(tableMetadata.sortKey, keys, tableMetadata.attributeDefinitions)
-    }
-
-    private fun getAttributesFromReturnValues(
-        returnValues: ReturnValue?,
-        prevItem: Map<String, AttributeValue>?
-    ): Map<String, AttributeValue>? =
-        when (returnValues) {
-            ReturnValue.ALL_OLD -> prevItem
-            ReturnValue.NONE, null -> null
-            else -> throw dynamoException("Return values set to invalid value")
-        }
-
-    private fun checkNumOfKeys(tableName: String, keys: Map<String, AttributeValue>) {
-        val tableMetadata = tablesMetadata[tableName] ?: throw ResourceNotFoundException.builder()
-            .message("Cannot do operations on a non-existent table").build()
-        val actualSize = if (tableMetadata.sortKey != null) 2 else 1
-        if (keys.size > actualSize) {
-            throw dynamoException("The number of conditions on the keys is invalid")
-        }
-    }
-
-    // TODO take into account that it's impossible to use expression-like and not-expression-like at the same query
-    private fun SelectRequest.select(): SelectResponse {
-        dynamoRequires(limit == null || limit > 0) {
-            "Limit must be >= 1."
-        }
-        dynamoRequires(consistentRead != false) {
-            "Non-consistent read is not supported in query yet."
-        }
-
-        dynamoRequires(tableName in tablesMetadata) {
-            "Cannot query from non-existent table $tableName."
-        }
-
-        val table = tablesMetadata.getValue(tableName)
-        val index = indexName?.let { table.localSecondaryIndex(it) }
-        val keyConditions = retrieveKeyConditions()?.also { keyConditions ->
-            keyConditions[table.partitionKey].let {
-                dynamoRequires(it != null && it.comparisonOperator() == ComparisonOperator.EQ) {
-                    "Partition key must use '=' operator."
-                }
-            }
-            dynamoRequires(keyConditions.size <= 2) {
-                "Length of key conditions must be 1 or 2."
-            }
-            if (index == null) {
-                dynamoRequires(keyConditions.size == 1 || table.sortKey in keyConditions) {
-                    "Only partition and sort keys can be used in key conditions."
-                }
-            } else {
-                dynamoRequires(keyConditions.size == 1 || index.sortKey in keyConditions) {
-                    "Only partition key and index sort key can be used in key conditions."
-                }
-            }
-        }
-
-        val items = storage.query(tableName, keyConditions).let {
-            if (scanIndexForward) it else it.reversed()
-        }.map { item ->
-            item.associate { it.name to it.type.toAttributeValue() }
-        }.let { items ->
-            exclusiveStartKey?.let { exclusiveStartKey ->
-                val sortKey = exclusiveStartKey[table.sortKey]
-                items.asSequence()
-                    .dropWhile { it[table.sortKey] != sortKey }
-                    .drop(1)
-                    .toList()
-            } ?: items
-        }.let { items ->
-            limit?.let { items.take(it) } ?: items
-        }
-
-        val filterExpression = retrieveFilterExpression()
-        val filteredItems = filterExpression?.let {
-            items.filter {
-                val withoutKeys = it.toMutableMap().apply {
-                    remove(table.partitionKey)
-                    remove(table.sortKey)
-                }
-                filterExpression.evaluate(withoutKeys)
-            }
-        } ?: items
-
-        val responseBuilder = SelectResponse.Builder().apply {
-            count = filteredItems.size
-            scannedCount = items.size
-        }
-
-        if (items.size == limit) {
-            responseBuilder.lastEvaluatedKey = items.last()
-        } else {
-            responseBuilder.lastEvaluatedKey = DefaultSdkAutoConstructMap.getInstance()
-        }
-
-        val transformer = retrieveAttributesTransformer(table)
-        if (select == Select.COUNT) {
-            responseBuilder.items = DefaultSdkAutoConstructList.getInstance()
-        } else {
-            responseBuilder.items = filteredItems.map(transformer)
-        }
-        return responseBuilder.build()
-    }
 }
